+++ .changelog/unreleased/breaking-changes/470-standalone-commitment-computation.md
@@ -1,0 +1,3 @@
+- Export standalone commitment computations for `PacketCommitment` and
+  `AcknowledgementCommitment`
+  ([#470](https://github.com/cosmos/ibc-rs/issues/470))
