+++ crates/ibc/src/core/ics04_channel/handler/recv_packet.rs
@@ -1,5 +1,6 @@
 use crate::core::ics03_connection::connection::State as ConnectionState;
 use crate::core::ics04_channel::channel::{Counterparty, Order, State};
+use crate::core::ics04_channel::commitment::compute::compute_packet_commitment;
 use crate::core::ics04_channel::error::ChannelError;
 use crate::core::ics04_channel::error::PacketError;
 use crate::core::ics04_channel::msgs::recv_packet::MsgRecvPacket;
@@ -79,7 +80,7 @@ where
             ClientConsensusStatePath::new(client_id_on_b, &msg.proof_height_on_a);
         let consensus_state_of_a_on_b = ctx_b.consensus_state(&client_cons_state_path_on_b)?;
 
-        let expected_commitment_on_a = ctx_b.compute_packet_commitment(
+        let expected_commitment_on_a = compute_packet_commitment(
             &msg.packet.data,
             &msg.packet.timeout_height_on_b,
             &msg.packet.timeout_timestamp_on_b,
