+++ crates/ibc/src/core/ics04_channel/handler/timeout_on_close.rs
@@ -1,5 +1,6 @@
 use crate::core::ics04_channel::channel::State;
 use crate::core::ics04_channel::channel::{ChannelEnd, Counterparty, Order};
+use crate::core::ics04_channel::commitment::compute_packet_commitment;
 use crate::core::ics04_channel::error::{ChannelError, PacketError};
 use crate::core::ics04_channel::msgs::timeout_on_close::MsgTimeoutOnClose;
 use crate::core::ics24_host::path::{
@@ -44,7 +45,7 @@ where
         Err(_) => return Ok(()),
     };
 
-    let expected_commitment_on_a = ctx_a.compute_packet_commitment(
+    let expected_commitment_on_a = compute_packet_commitment(
         &packet.data,
         &packet.timeout_height_on_b,
         &packet.timeout_timestamp_on_b,
@@ -161,9 +162,9 @@ where
 
 #[cfg(test)]
 mod tests {
+    use crate::core::ics04_channel::commitment::compute_packet_commitment;
     use crate::core::ics04_channel::commitment::PacketCommitment;
     use crate::core::ics04_channel::handler::timeout_on_close::validate;
-    use crate::core::ValidationContext;
     use crate::mock::context::MockContext;
     use crate::prelude::*;
     use crate::Height;
@@ -204,7 +205,7 @@ mod tests {
 
         let packet = msg.packet.clone();
 
-        let packet_commitment = context.compute_packet_commitment(
+        let packet_commitment = compute_packet_commitment(
             &msg.packet.data,
             &msg.packet.timeout_height_on_b,
             &msg.packet.timeout_timestamp_on_b,
