+++ crates/ibc/src/test_utils.rs
@@ -325,32 +325,6 @@ impl SendPacketValidationContext for DummyTransferModule {
             .into()),
         }
     }
-
-    fn hash(&self, value: &[u8]) -> Vec<u8> {
-        use sha2::Digest;
-
-        sha2::Sha256::digest(value).to_vec()
-    }
-
-    fn compute_packet_commitment(
-        &self,
-        packet_data: &[u8],
-        timeout_height: &crate::core::ics04_channel::timeout::TimeoutHeight,
-        timeout_timestamp: &crate::timestamp::Timestamp,
-    ) -> PacketCommitment {
-        let mut hash_input = timeout_timestamp.nanoseconds().to_be_bytes().to_vec();
-
-        let revision_number = timeout_height.commitment_revision_number().to_be_bytes();
-        hash_input.append(&mut revision_number.to_vec());
-
-        let revision_height = timeout_height.commitment_revision_height().to_be_bytes();
-        hash_input.append(&mut revision_height.to_vec());
-
-        let packet_data_hash = self.hash(packet_data);
-        hash_input.append(&mut packet_data_hash.to_vec());
-
-        self.hash(&hash_input).into()
-    }
 }
 
 impl SendPacketExecutionContext for DummyTransferModule {
